backend/CIMAS/requirements.txt
asgiref==3.9.1
Django==5.2.6
django-cors-headers==4.9.0
django-seed==0.3.1
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
Faker==37.8.0
pillow==11.3.0
psycopg2-binary==2.9.10
PyJWT==2.10.1
python-dotenv==1.1.1
sqlparse==0.5.3
toposort==1.10
tzdata==2025.2
